+++ pkg/auth/service_test.go
@@ -17,7 +17,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
-	"github.com/treeverse/lakefs/contrib/auth-acl"
+	authacl "github.com/treeverse/lakefs/contrib/auth/acl"
 	"github.com/treeverse/lakefs/pkg/auth"
 	"github.com/treeverse/lakefs/pkg/auth/acl"
 	"github.com/treeverse/lakefs/pkg/auth/crypt"
@@ -108,7 +108,7 @@ func userWithACLs(t testing.TB, s auth.Service, a model.ACL) string {
 func TestAuthService_ListUsers_PagedWithPrefix(t *testing.T) {
 	ctx := context.Background()
 	kvStore := kvtest.GetStore(ctx, t)
-	s := auth_acl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
+	s := authacl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
 		Enabled: false,
 	}, logging.ContextUnavailable())
 
@@ -158,7 +158,7 @@ func TestAuthService_ListUsers_PagedWithPrefix(t *testing.T) {
 func TestAuthService_ListPaged(t *testing.T) {
 	ctx := context.Background()
 	kvStore := kvtest.GetStore(ctx, t)
-	s := auth_acl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
+	s := authacl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
 		Enabled: false,
 	}, logging.ContextUnavailable())
 
@@ -541,16 +541,16 @@ func BenchmarkKVAuthService_ListEffectivePolicies(b *testing.B) {
 	ctx := context.Background()
 	kvStore := kvtest.GetStore(ctx, b)
 
-	serviceWithoutCache := auth_acl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
+	serviceWithoutCache := authacl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
 		Enabled: false,
 	}, logging.ContextUnavailable())
-	serviceWithCache := auth_acl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
+	serviceWithCache := authacl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
 		Enabled: true,
 		Size:    1024,
 		TTL:     20 * time.Second,
 		Jitter:  3 * time.Second,
 	}, logging.ContextUnavailable())
-	serviceWithCacheLowTTL := auth_acl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
+	serviceWithCacheLowTTL := authacl.NewAuthService(kvStore, crypt.NewSecretStore(someSecret), authparams.ServiceCache{
 		Enabled: true,
 		Size:    1024,
 		TTL:     1 * time.Millisecond,
@@ -569,7 +569,7 @@ func BenchmarkKVAuthService_ListEffectivePolicies(b *testing.B) {
 	})
 }
 
-func benchmarkKVListEffectivePolicies(b *testing.B, s *auth_acl.AuthService, userName string) {
+func benchmarkKVListEffectivePolicies(b *testing.B, s *authacl.AuthService, userName string) {
 	b.ResetTimer()
 	ctx := context.Background()
 	for n := 0; n < b.N; n++ {
